# design/api-spec.yaml
openapi: 3.0.1
info:
  title: "hicto"
  description: "figure-based event representation"
  version: v1
servers:
  - url: /main-api
paths:
  /api/v1/event:
    post:
      tags:
        - 'Hicto'
      operationId: createEvent
      responses:
        '200':
          description: "Success"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/eventDto'
                nullable: true 

  /api/v1/event/{event_id}:
    get: 
      tags:
        - 'Hicto'
      operationId: getEvent
      parameters:
        - in: path
          name: event_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: "Success"
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/eventDto'
                nullable: true
    put:
      tags:
        - 'Hicto'
      operationId: updateEvent
      parameters:
        - in: path
          name: event_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: "Success"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/eventDto'
                nullable: true
    delete:
      tags:
        - 'Hicto'
      operationId: deleteEvent
      parameters:
        - in: path
          name: event_id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: "Success"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/eventDto'
                nullable: true
components:
  schemas:
    eventDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1007
        title:
          type: string
          example: "Bose-Einstein particle"
        contents:
          type: string
          example: "Bose in India..."
        updatedAt:
          type: string
          format: date-time
          example: "2022-06-17T06:01:43+09:00"
        createdAt:
          type: string
          format: date-time
          example: "2022-06-17T06:05:43+09:00"
          
        
        